refactor docx to separated files
